feat: "회원탈퇴&찜 리스트&주문목록&주문목록,취소내역 날짜 조회 Controller"

+++ src/main/java/controller/WithdrawalAction.java
@@ -11,17 +11,22 @@ public class WithdrawalAction implements Action {
 	@Override
 	public ActionForward execute(HttpServletRequest request, HttpServletResponse response) throws Exception {
 		ActionForward forward = new ActionForward();
+	
 		forward.setPath("main.do");
 		forward.setRedirect(true);
-		
+
 		MemberDAO mdao = new MemberDAO();
 		MemberVO mvo = new MemberVO();
-		String mId = (String)request.getSession().getAttribute("mId");
-		mvo.setmId(mId);
-		
-		if (mdao.delete(mvo)) {
-			request.getSession().invalidate();
-			return forward;
+		int mNum = (int)request.getSession().getAttribute("mId");
+		mvo.setmPw(request.getParameter("mPw"));
+		mvo.setmNum(mNum);
+
+		mvo=mdao.selectOneLogin(mvo);//탈퇴하려는 id랑 pw랑 같고 그게 멤버 테이블에 있으면
+		if(mvo !=null ) {
+			if (mdao.delete(mvo)) {
+				request.getSession().invalidate();
+				return forward;
+			}
 		}
 		return null;
 	}
